misc/HtmlTestSuite/results/test-1144.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head>

<body>
<script>
  var t = async_test("Test that the child iframe navigation is allowed");
  window.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data.result, 'success');
  });
</script>
<iframe src="support/link_click_navigation.sub.html?csp=navigate-to%20http%3A%2F%2F{{domains[www1]}}:{{ports[http][0]}}&amp;target=http%3A%2F%2F{{domains[www1]}}:{{ports[http][0]}}%2Fcontent-security-policy%2Fnavigate-to%2Fsupport%2Fpost_message_to_frame_owner.html">
&lt;/body&gt;</iframe></body></html>
